chore: cursor should be size_t

+++ check.c
@@ -4,7 +4,7 @@
 
 pgn_check_t __pgn_check_from_string(char *str, size_t *consumed)
 {
-    int cursor = 0;
+    size_t cursor = 0;
     pgn_check_t check = PGN_CHECK_NONE;
 
     while (str[cursor] == '+') {
